+++ arc/abc/option.py
@@ -24,10 +24,12 @@ __all__ = (
     "OptionWithChoicesParams",
     "OptionBase",
     "CommandOptionBase",
+    "ConverterOption",
     "OptionType",
 )
 
 T = t.TypeVar("T")
+OriginT = t.TypeVar("OriginT")
 
 Option = t.Annotated
 """Alias for typing.Annotated.
@@ -92,6 +94,14 @@ class OptionType(enum.IntEnum):
     ATTACHMENT = 11
     """Denotes a command option where the value will be an attachment."""
 
+    # Custom optiontypes are 10k+
+
+    MEMBER = 10001
+    """Denotes a command option where the value will be resolved to a member."""
+
+    COLOR = 10002
+    """Denotes a command option where the value will be a color."""
+
     @classmethod
     def from_hikari(cls, option_type: hikari.OptionType) -> OptionType:
         """Convert a hikari.OptionType to an OptionType."""
@@ -99,10 +109,13 @@ class OptionType(enum.IntEnum):
 
     def to_hikari(self) -> hikari.OptionType:
         """Convert an OptionType to a hikari.OptionType."""
-        # TODO: Map custom option types to their respective hikari.OptionType
-        return hikari.OptionType(self.value)
+        if self.value < 10000:
+            return hikari.OptionType(self.value)
 
-    # TODO: When adding custom convertible option types, add them with an offset of 1000 or so
+        if self is OptionType.MEMBER:
+            return hikari.OptionType.USER
+        else:
+            return hikari.OptionType.STRING
 
 
 class OptionParams(t.Generic[T]):
@@ -299,6 +312,15 @@ class CommandOptionBase(OptionBase[T], t.Generic[T, ClientT, ParamsT]):
 
     def _to_dict(self) -> dict[str, Any]:
         return {**super()._to_dict(), "is_required": self.is_required}
+
+
+@attr.define(slots=True, kw_only=True)
+class ConverterOption(CommandOptionBase[T, ClientT, ParamsT], t.Generic[T, ClientT, ParamsT, OriginT]):
+    """An option with a built-in converter."""
+
+    @abc.abstractmethod
+    def _convert_value(self, value: OriginT) -> T:
+        """Convert the value to the desired type."""
 
 
 @attr.define(slots=True, kw_only=True)