+++ arc/command/option/__init__.py
@@ -1,6 +1,7 @@
 from .attachment import AttachmentOption, AttachmentParams
 from .bool import BoolOption, BoolParams
 from .channel import ChannelOption, ChannelParams
+from .custom import ColorOption, ColorParams, ColourOption, ColourParams, MemberOption, MemberParams
 from .float import FloatOption, FloatParams
 from .int import IntOption, IntParams
 from .mentionable import MentionableOption, MentionableParams
@@ -27,6 +28,12 @@ __all__ = (
     "MentionableParams",
     "AttachmentOption",
     "AttachmentParams",
+    "MemberOption",
+    "MemberParams",
+    "ColorOption",
+    "ColorParams",
+    "ColourOption",
+    "ColourParams",
 )
 
 # MIT License